게시판 기능완료

+++ src/main/java/com/example/teamprojectbringiton/board/BoardController.java
@@ -1,5 +1,7 @@
 package com.example.teamprojectbringiton.board;
 
+import com.example.teamprojectbringiton._core.handler.exception.CustomRestfullException;
+import com.example.teamprojectbringiton._core.handler.exception.UnAuthorizedException;
 import com.example.teamprojectbringiton.board.dto.request.BoardUpdateDTO;
 import com.example.teamprojectbringiton.board.dto.request.BoardWriteDTO;
 import com.example.teamprojectbringiton.board.dto.response.BoardDTO;
@@ -11,6 +13,7 @@ import jakarta.servlet.http.HttpSession;
 import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,6 +31,9 @@ public class BoardController {
 
     @Autowired
     BoardService boardService;
+
+    @Autowired
+    BoardRepository boardRepository;
 
     @GetMapping("/board-main")
     public String boardPage(Model model){
@@ -82,11 +88,23 @@ public class BoardController {
     @PostMapping("/board-write")
     public String boardWrite(BoardWriteDTO dto) {
         User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new UnAuthorizedException("로그인 해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
         System.out.println("1111@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@" + sessionUser.getId());
         dto.setUserId(sessionUser.getId());
         // HTML 태그 제거
         String plainText = Jsoup.clean(dto.getBoardContent(), Whitelist.none());
         dto.setBoardContent(plainText);
+        if (dto.getBoardTitle() == null || dto.getBoardTitle().isEmpty()) {
+            throw new CustomRestfullException("댓글을 입력해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
+        if (dto.getBoardContent() == null || dto.getBoardContent().isEmpty()) {
+            throw new CustomRestfullException("댓글을 입력해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
         boardService.boardWrite(dto);
         System.out.println("2222@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@" + sessionUser.getId());
         return "redirect:/board-main";
@@ -95,7 +113,16 @@ public class BoardController {
 
     @GetMapping("/board-update/{id}")
     public String boardUpdatePage(@PathVariable Integer id, Model model){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new UnAuthorizedException("로그인 해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
         BoardDTO boardDTO = boardService.findById(id);
+        if (boardDTO.getUserId() != sessionUser.getId()) {
+            throw new CustomRestfullException("권한이 없습니다",
+                    HttpStatus.BAD_REQUEST);
+        }
         model.addAttribute("boardDTO", boardDTO);
         return "board/boardUpdatePage";
     }
@@ -103,6 +130,18 @@ public class BoardController {
     @PostMapping("/board-update/{id}")
     public String boardUpdate(BoardUpdateDTO dto){
         User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new UnAuthorizedException("로그인 해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
+        if (dto.getBoardTitle() == null || dto.getBoardTitle().isEmpty()) {
+            throw new CustomRestfullException("제목을 입력해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
+        if (dto.getBoardContent() == null || dto.getBoardContent().isEmpty()) {
+            throw new CustomRestfullException("내용을 입력해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
         dto.setUserId(sessionUser.getId());
         boardService.boardUpdate(dto, dto.getId());
         int boardUpdateId = dto.getId();
@@ -112,6 +151,16 @@ public class BoardController {
 
     @GetMapping("/board-delete/{id}")
     public String boardDelete(@PathVariable Integer id){
+        User sessionUser = (User) session.getAttribute("sessionUser");
+        if (sessionUser == null) {
+            throw new UnAuthorizedException("로그인 해주세요",
+                    HttpStatus.BAD_REQUEST);
+        }
+        BoardDTO boardDTO = boardRepository.findById(id);
+        if (boardDTO.getUserId() != sessionUser.getId()) {
+            throw new CustomRestfullException("권한이 없습니다",
+                    HttpStatus.BAD_REQUEST);
+        }
         boardService.deleteById(id);
         return "redirect:/board-main";
     }
